Add logging to Telnyx webhook handler
