+++ src/activities/Elsa.Activities.Telnyx/Webhooks/Handlers/TriggerFilteredCallInitiatedActivities.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 using Elsa.Activities.Telnyx.Activities;
-using Elsa.Activities.Telnyx.Models;
+using Elsa.Activities.Telnyx.Extensions;
 using Elsa.Activities.Telnyx.Providers.Bookmarks;
 using Elsa.Activities.Telnyx.Webhooks.Events;
-using Elsa.Activities.Telnyx.Webhooks.Payloads.Abstract;
 using Elsa.Activities.Telnyx.Webhooks.Payloads.Call;
 using Elsa.Models;
 using Elsa.Services;
@@ -33,7 +31,7 @@ namespace Elsa.Activities.Telnyx.Webhooks.Handlers
             if (payload.Direction != "incoming")
                 return;
             
-            var correlationId = GetCorrelationId(payload);
+            var correlationId = payload.GetCorrelationId();
             var toBookmark = new FilteredCallInitiatedToBookmark(payload.To);
             var toQuery = new WorkflowsQuery(nameof(FilteredCallInitiated), toBookmark, correlationId);
             var fromBookmark = new FilteredCallInitiatedFromBookmark(payload.From);
@@ -41,20 +39,6 @@ namespace Elsa.Activities.Telnyx.Webhooks.Handlers
 
             await _workflowLaunchpad.CollectAndDispatchWorkflowsAsync(toQuery, new WorkflowInput(webhook), cancellationToken);
             await _workflowLaunchpad.CollectAndDispatchWorkflowsAsync(fromQuery, new WorkflowInput(webhook), cancellationToken);
-        }
-        
-        private string GetCorrelationId(Payload payload)
-        {
-            if (!string.IsNullOrWhiteSpace(payload.ClientState))
-            {
-                var clientStatePayload = ClientStatePayload.FromBase64(payload.ClientState);
-                return clientStatePayload.CorrelationId;
-            }
-
-            if (payload is CallPayload callPayload)
-                return callPayload.CallSessionId;
-
-            throw new NotSupportedException($"The received payload type {payload.GetType().Name} is not supported yet.");
         }
     }
 }
